% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{ACAT_Pval}
\alias{ACAT_Pval}
\title{Cauchy combination of Pvalues from individual association tests.}
\usage{
ACAT_Pval(Zs)
}
\arguments{
\item{Zs}{The matrix of Z scores (rows are SNPs and column are traits)}
}
\value{
A numeric vector of Pvalues rowwise(ie SNP)
}
\description{
Cauchy combination of Pvalues from individual association tests.
}
\examples{
\dontrun{
Zs<-sapply(1:5,function(x){rnorm(10)})
ACAT_Pval(Zs)}
}
