% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Z-score Matrix}
\format{
A numeric matrix with 10000 rows (SNPs) and 40 columns (traits).
}
\usage{
data(example_data)
}
\description{
A simulated matrix of Z-scores for testing multi-phenotype methods. The matrix was
Generated for 40 phenotypes.
}
\examples{
data(example_data)
str(example_data)
}
\keyword{datasets}
