% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityDetection.R
\name{get_n_cluster}
\alias{get_n_cluster}
\title{Determine Optimal Number of Clusters Using Modularity}
\usage{
get_n_cluster(hc, Sigma, m = ncol(Sigma), between_cluster = 0.8)
}
\arguments{
\item{hc}{A hierarchical clustering object (from \code{hclust()}).}

\item{Sigma}{A simialrity matrix used for clustering (e.g., correlation or partial correlation matrix).}

\item{m}{Maximum number of clusters to evaluate. Default is number of columns in Sigma.}

\item{between_cluster}{Threshold: if all entries in Sigma > threshold, returns 1 cluster}
}
\value{
A list with:
-- \code{n_cluster}: optimal number of clusters
--\code{Qmodularity}: modularity scores for each tested cluster number
}
\description{
Computes the optimal number of clusters in hierarchical clustering
using a modularity-based criterion.Maximizing modularity
}
\examples{
set.seed(123)
K<-6
R=matrix(0.5,nrow=K,ncol=K)
diag(R)<-1
hc<-hclust(as.dist(1-R))
res<-get_n_cluster(hc,R)
}
