% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{HCLC_Pval}
\alias{HCLC_Pval}
\title{HCLC Test for Combining Z-scores across Correlated Traits}
\usage{
HCLC_Pval(Zs)
}
\arguments{
\item{Zs}{The matrix of Z scores (rows are SNPs and column are traits)}
}
\value{
A numeric vector of pvalues row-wise (SNP wise)
}
\description{
Computes the OBrien test p-values for rows of Z-scores, accounting for correlation among traits.
}
\examples{
\dontrun{Zs<-sapply(1:5,function(x){rnorm(10)})
HCLC_Pval(Zs)}
}
