% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{OBrien_Pval}
\alias{OBrien_Pval}
\title{OBrien test for testing an association between SNPs and multiple correlated traits.
Computes the OBrien test p-values for rows of Z-scores matrix , accounting for correlation among traits.}
\usage{
OBrien_Pval(Zs)
}
\arguments{
\item{Zs}{The matrix of Z scores (rows are SNPs and column are traits)}
}
\value{
A numeric vector of pvalues row-wise (SNP)
}
\description{
OBrien test for testing an association between SNPs and multiple correlated traits.
Computes the OBrien test p-values for rows of Z-scores matrix , accounting for correlation among traits.
}
\examples{
\dontrun{Zs<-sapply(1:5,function(x){rnorm(10)})
OBrien_Pval(Zs)}
}
