% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{PCFisher_Pval}
\alias{PCFisher_Pval}
\title{PCFisher Test for Combining Z-scores across Correlated Traits}
\usage{
PCFisher_Pval(Zs)
}
\arguments{
\item{Zs}{a Matrix of Z scores (rows are SNPs and columns are traits)}
}
\value{
A numeric vector of pvalues for each row (i.e SNPs)
}
\description{
Computes the OBrien test p-values for rows of Z-scores, accounting for correlation among traits.
}
\examples{
\dontrun{Zs<-sapply(1:5,function(x){rnorm(10)})
PCFisher_Pval(Zs)}
}
