% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypothesis_Tests.R
\name{Test_With_Network}
\alias{Test_With_Network}
\title{Function to run the Multiple Phenotype Association Test}
\usage{
Test_With_Network(Zs, B, method = "ACAT")
}
\arguments{
\item{Zs}{The Matrix of Z scores (M by K) M is SNPs and K is traits}

\item{B}{Matrix of Community structure (K by L)}

\item{method}{The name of multiple phenotype test method. The default is ACAT.}
}
\value{
The Matrix of Pvalues, where each column is the Pvalues from each cluster
}
\description{
Function to run the Multiple Phenotype Association Test
}
\examples{
#load the example data
data<-example_data
Cluster<-Get_Clusters(data)
pval_out<-Test_With_Network(Zs=data,B=Cluster$B_cor,method="ACAT")
head(pval_out)
}
