% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypothesis_Tests.R
\name{Test_Without_Network}
\alias{Test_Without_Network}
\title{Run an association test of SNPs with multiple phenotype without considering network modules}
\usage{
Test_Without_Network(Zs, method = "ACAT")
}
\arguments{
\item{Zs}{The matrix of Z scores; rows are SNPs and columns are traits}

\item{method}{method is the test method, There are five methods to check and default is ACAT
We plan to include other methods in the future}
}
\value{
A numeric vector of pvalues of length equal to number of SNPs
}
\description{
Run an association test of SNPs with multiple phenotype without considering network modules
}
\examples{
#load the example data
data<-example_data
Cluster<-Get_Clusters(data)
pval_out<-Test_With_Network(Zs=data,B=Cluster$B_cor,method="ACAT")
head(pval_out)

}
