% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{Wald_Pval}
\alias{Wald_Pval}
\title{Wald Test for Combining Z-scores across Correlated Traits}
\usage{
Wald_Pval(Zs)
}
\arguments{
\item{Zs}{The matrix of Z scores (rows are SNPs and column are traits)}
}
\value{
A numeric vector vector of pvalues rowwise (per SNP)
}
\description{
Computes the OBrien test p-values for rows of Z-scores, accounting for correlation among traits.
}
\examples{
\dontrun{Zs<-sapply(1:5,function(x){rnorm(10)})
Wald_Pval(Zs)}
}
