% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityDetection.R
\name{Get_Clusters}
\alias{Get_Clusters}
\title{Get Clusters from Correlation and Partial Correlation Matrices
Constructs clustering structures from both the correlation matrix and the
partial correlation matrix estimated using sparse GGM.}
\usage{
Get_Clusters(Zscores, penalty_function = "lasso")
}
\arguments{
\item{Zscores}{A numeric matrix of Z-scores (rows = SNPs, columns = traits).}

\item{penalty_function}{Penalty type for GGMncv (default: "lasso")}
}
\value{
A list with:
B_cor\verb{: Cluster matrix from correlation-based clustering B_Pcor}: Cluster matrix from partial correlation-based clustering
}
\description{
Get Clusters from Correlation and Partial Correlation Matrices
Constructs clustering structures from both the correlation matrix and the
partial correlation matrix estimated using sparse GGM.
}
\examples{
Z <- MASS::mvrnorm(100, mu = rep(0, 10), Sigma = diag(10))
res <- Get_Clusters(Z)
}
